# BiBoGC/openapi-bibogc.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BiBo's Grocery Store API",
    "description": "API cho hệ thống quản lý cửa hàng tạp hoá BiBo. Bao gồm quản lý sản phẩm, danh mục, nhà cung cấp, lô hàng và giao dịch kho.\n\n## Tính năng chính:\n- **Products**: Quản lý sản phẩm (CRUD)\n- **Categories**: Quản lý danh mục sản phẩm\n- **Suppliers**: Quản lý nhà cung cấp\n- **Product Batches**: Quản lý lô hàng (theo dõi hạn sử dụng)\n- **Stock Transactions**: Ghi nhận giao dịch kho (nhập/xuất/điều chỉnh)\n\n## Quy tắc chung:\n- Tất cả ID đều là GUID\n- Pagination sử dụng pageNumber (1-based) và pageSize\n- Soft delete: Dữ liệu bị xóa vẫn được lưu trong DB\n- Validation errors trả về 400 Bad Request với chi tiết lỗi",
    "version": "1.0.0",
    "contact": {
      "name": "BiBo's GC Support",
      "email": "support@bibogc.local"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5020",
      "description": "Development Server"
    }
  ],
...
  "paths": {
...
    "/api/suppliers": {
      "get": {
        "tags": ["Suppliers"],
        "summary": "Lấy danh sách nhà cung cấp (phân trang)",
        "description": "Lấy danh sách nhà cung cấp với hỗ trợ phân trang, tìm kiếm và lọc.\n\n**Tìm kiếm:**\n- Tìm theo tên, người liên hệ, số điện thoại, địa chỉ\n- Không phân biệt hoa thường (ILike)",
        "operationId": "getSuppliers",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "description": "Số trang",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Số lượng mỗi trang",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            }
          },
          {
            "name": "searchTerm",
            "in": "query",
            "description": "Từ khóa tìm kiếm",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "in": "query",
            "description": "Lọc theo trạng thái hoạt động (null = tất cả)",
            "schema": {
              "type": "boolean",
              "default": true
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Sắp xếp theo field",
            "schema": {
              "type": "string",
              "enum": ["Name", "CreatedAt"],
              "default": "Name"
            }
          },
          {
            "name": "sortDescending",
            "in": "query",
            "description": "Sắp xếp giảm dần",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách nhà cung cấp",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedSupplierResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Suppliers"],
        "summary": "Tạo nhà cung cấp mới",
        "description": "Tạo một nhà cung cấp mới.\n\n**Validation:**\n- name: Bắt buộc, tối đa 200 ký tự\n- contactPerson: Tùy chọn, tối đa 100 ký tự\n- phoneNumber: Tùy chọn, tối đa 20 ký tự\n- address: Tùy chọn, tối đa 500 ký tự",
        "operationId": "createSupplier",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupplierCommand"
              },
              "example": {
                "name": "Công ty TNHH Coca-Cola Việt Nam",
                "contactPerson": "Nguyễn Văn An",
                "phoneNumber": "0283456789",
                "address": "KCN Biên Hòa, Đồng Nai"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Nhà cung cấp đã được tạo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierDto"
                }
              }
            }
          },
          "400": {
            "description": "Lỗi validation hoặc tên đã tồn tại",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/suppliers/{id}": {
      "get": {
        "tags": ["Suppliers"],
        "summary": "Lấy chi tiết nhà cung cấp",
        "description": "Lấy thông tin chi tiết nhà cung cấp bao gồm 5 giao dịch gần nhất.",
        "operationId": "getSupplier",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID nhà cung cấp (GUID)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chi tiết nhà cung cấp",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierDto"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy nhà cung cấp",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Suppliers"],
        "summary": "Cập nhật nhà cung cấp",
        "description": "Cập nhật thông tin nhà cung cấp.",
        "operationId": "updateSupplier",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID nhà cung cấp cần cập nhật",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSupplierCommand"
              },
              "example": {
                "name": "Công ty TNHH Coca-Cola VN",
                "contactName": "Trần Văn B",
                "contactPhone": "0909123456",
                "address": "KCN Biên Hòa 2, Đồng Nai",
                "isActive": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Nhà cung cấp đã được cập nhật",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierDto"
                }
              }
            }
          },
          "400": {
            "description": "Lỗi validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy nhà cung cấp",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Suppliers"],
        "summary": "Xóa nhà cung cấp (soft delete)",
        "description": "Xóa mềm nhà cung cấp.",
        "operationId": "deleteSupplier",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID nhà cung cấp cần xóa",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Nhà cung cấp đã được xóa"
          },
          "404": {
            "description": "Không tìm thấy nhà cung cấp",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/suppliers/{id}/activate": {
      "post": {
        "tags": ["Suppliers"],
        "summary": "Kích hoạt nhà cung cấp",
        "description": "Đặt trạng thái nhà cung cấp thành hoạt động (isActive = true).",
        "operationId": "activateSupplier",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID nhà cung cấp",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Nhà cung cấp đã được kích hoạt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierDto"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy nhà cung cấp",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/suppliers/{id}/deactivate": {
      "post": {
        "tags": ["Suppliers"],
        "summary": "Vô hiệu hóa nhà cung cấp",
        "description": "Đặt trạng thái nhà cung cấp thành không hoạt động (isActive = false).",
        "operationId": "deactivateSupplier",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID nhà cung cấp",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Nhà cung cấp đã được vô hiệu hóa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierDto"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy nhà cung cấp",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SupplierDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "ID nhà cung cấp"
          },
          "name": {
            "type": "string",
            "description": "Tên nhà cung cấp"
          },
          "contactName": {
            "type": "string",
            "nullable": true,
            "description": "Người liên hệ"
          },
          "contactPhone": {
            "type": "string",
            "nullable": true,
            "description": "Số điện thoại"
          },
          "address": {
            "type": "string",
            "nullable": true,
            "description": "Địa chỉ"
          },
          "isActive": {
            "type": "boolean",
            "description": "Trạng thái hoạt động"
          },
          "createAt": {
            "type": "string",
            "format": "date-time",
            "description": "Ngày tạo"
          },
          "transactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StockTransactionDto"
            },
            "nullable": true,
            "description": "Danh sách giao dịch gần đây"
          }
        }
      },
      "SuppliersDto": {
        "type": "object",
        "description": "DTO cho danh sách NCC (không có transactions)",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "contactName": {
            "type": "string",
            "nullable": true
          },
          "contactPhone": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "CreateSupplierCommand": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "description": "Tên nhà cung cấp (bắt buộc)"
          },
          "contactPerson": {
            "type": "string",
            "maxLength": 100,
            "nullable": true,
            "description": "Người liên hệ"
          },
          "phoneNumber": {
            "type": "string",
            "maxLength": 20,
            "nullable": true,
            "description": "Số điện thoại"
          },
          "address": {
            "type": "string",
            "maxLength": 500,
            "nullable": true,
            "description": "Địa chỉ"
          }
        }
      },
      "UpdateSupplierCommand": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "description": "Tên nhà cung cấp"
          },
          "contactName": {
            "type": "string",
            "maxLength": 100,
            "nullable": true,
            "description": "Người liên hệ"
          },
          "contactPhone": {
            "type": "string",
            "maxLength": 20,
            "nullable": true,
            "description": "Số điện thoại"
          },
          "address": {
            "type": "string",
            "maxLength": 500,
            "nullable": true,
            "description": "Địa chỉ"
          },
          "isActive": {
            "type": "boolean",
            "description": "Trạng thái hoạt động"
          }
        }
      },
      "StockTransactionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "ID giao dịch"
          },
          "productId": {
            "type": "string",
            "format": "uuid",
            "description": "ID sản phẩm"
          },
          "productName": {
            "type": "string",
            "description": "Tên sản phẩm"
          },
          "productBatchId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "ID lô hàng"
          },
          "batchNumber": {
            "type": "string",
            "nullable": true,
            "description": "Số lô"
          },
          "supplierId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "ID nhà cung cấp"
          },
          "supplierName": {
            "type": "string",
            "nullable": true,
            "description": "Tên nhà cung cấp"
          },
          "sku": {
            "type": "string",
            "description": "Mã SKU"
          },
          "transactionType": {
            "type": "string",
            "description": "Loại giao dịch (Purchase, Sale, AdjustmentIn, AdjustmentOut, Damage, Expiry, Return, SupplierReturn)"
          },
          "quantity": {
            "type": "integer",
            "description": "Số lượng"
          },
          "unitPrice": {
            "type": "number",
            "format": "decimal",
            "description": "Đơn giá"
          },
          "totalAmount": {
            "type": "number",
            "format": "decimal",
            "description": "Tổng tiền"
          },
          "transactionDate": {
            "type": "string",
            "format": "date-time",
            "description": "Ngày giao dịch"
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "description": "Ghi chú"
          },
          "referenceNumber": {
            "type": "string",
            "nullable": true,
            "description": "Số tham chiếu"
          }
        }
      },
...
      "PaginatedSupplierResult": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SuppliersDto"
            }
          },
          "pageNumber": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "totalCount": {
            "type": "integer"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          }
        }
      },
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  }
}